requirements.txt
alembic==1.14.0
psycopg2-binary==2.9.9
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-core==2.27.2
python-decouple==3.8
python-dotenv==1.0.1
requests>=2.25.0
SQLAlchemy==2.0.28
flask-sqlalchemy
flask==3.0.1
email-validator==2.2.0
pyjwt==2.10.0
confluent_kafka==2.7.0
pillow==11.1.0
redis==5.0.4
web3==7.7.0
eth-account==0.13.4
flask-socketio==5.5.1
eventlet==0.36.0
kafka-python==2.0.2
testcontainers[kafka]
sortedcontainers==2.4.0
urllib3==2.2.1
Werkzeug==3.0.1
gunicorn==22.0.0
asyncpg
flasgger
phonenumbers==9.0.9
hdwallet==2.2.1
ecdsa
eth-utils==5.0.0
bitcart==1.18.0.0
cryptography==43.0.3
tronpy==0.5.0
python-binance==1.0.21
grpcio-tools==1.73.1
pyotp==2.9.0
flask-admin
flask-login
solana
solders
blockcypher==1.0.93